browser: add configuration option for isolated install
