+++ browser/src/robocorp/browser/__init__.py
@@ -9,7 +9,7 @@ from playwright.sync_api import (
     Playwright,
 )
 
-from ._browser_engines import BrowserEngine
+from ._types import BrowserEngine, BrowserNotFound, InstallError
 
 __version__ = "2.1.0"
 version_info = [int(x) for x in __version__.split(".")]
@@ -37,9 +37,9 @@ def configure(**kwargs) -> None:
         See also: `robocorp.browser.configure_context` to change other
         arguments related to the browser context creation.
     """  # noqa
-    from ._browser_context import _browser_config
+    from ._config import browser_config
 
-    config = _browser_config()
+    config = browser_config()
 
     for key, value in kwargs.items():
         if key == "viewport_size":
@@ -74,9 +74,9 @@ def configure_context(**kwargs) -> None:
         to call `robocorp.browser.context(...)` with the required arguments
         directly.
     """
-    from . import _browser_context
+    from . import _context
 
-    browser_context_kwargs = _browser_context.browser_context_kwargs()
+    browser_context_kwargs = _context.browser_context_kwargs()
     browser_context_kwargs.update(kwargs)
 
 
@@ -97,9 +97,9 @@ def page() -> Page:
         page = browser.context().new_page()
         ```
     """
-    from . import _browser_context
+    from . import _context
 
-    return _browser_context.page()
+    return _context.page()
 
 
 def browser() -> Browser:
@@ -118,9 +118,9 @@ def browser() -> Browser:
         Note that the returned browser must not be closed. It will be
         automatically closed when the task run session finishes.
     """
-    from . import _browser_context
+    from . import _context
 
-    return _browser_context.browser()
+    return _context.browser()
 
 
 def playwright() -> Playwright:
@@ -139,9 +139,9 @@ def playwright() -> Playwright:
         Note that the returned instance must not be closed. It will be
         automatically closed when the task run session finishes.
     """
-    from . import _browser_context
+    from . import _context
 
-    return _browser_context.playwright()
+    return _context.playwright()
 
 
 def context(**kwargs) -> BrowserContext:
@@ -165,9 +165,9 @@ def context(**kwargs) -> BrowserContext:
         If the context was already previously created the **kwargs passed will
         be ignored.
     """
-    from . import _browser_context
+    from . import _context
 
-    return _browser_context.context(**kwargs)
+    return _context.context(**kwargs)
 
 
 def goto(url: str) -> Page:
@@ -211,9 +211,9 @@ def screenshot(
     from robocorp import log
 
     if element is None:
-        from . import _browser_context
+        from . import _context
 
-        element = _browser_context.page()
+        element = _context.page()
 
     with log.suppress():
         # Suppress log because we don't want the bytes to appear at
@@ -229,7 +229,7 @@ def screenshot(
         return in_bytes
 
 
-def install(browser_engine: BrowserEngine):
+def install(browser_engine: BrowserEngine, force: bool = False):
     """
     Downloads and installs the given browser engine.
 
@@ -240,9 +240,9 @@ def install(browser_engine: BrowserEngine):
     Args:
         browser_engine: Browser engine which should be installed
     """
-    from . import _browser_engines
+    from . import _engines
 
-    _browser_engines.install_browser(browser_engine, force=False)
+    _engines.install_browser(browser_engine, force=force)
 
 
 __all__ = [
@@ -256,4 +256,6 @@ __all__ = [
     "screenshot",
     "install",
     "BrowserEngine",
+    "InstallError",
+    "BrowserNotFound",
 ]
