+++ browser/src/robocorp/browser/cli.py
@@ -1,22 +1,39 @@
+# ruff: noqa: E501
 import argparse
 import sys
 
-from ._browser_engines import BrowserEngine, InstallError, install_browser
+from ._engines import BrowserEngine, InstallError, install_browser
 
 
 def main():
-    parser = argparse.ArgumentParser(prog="robocorp.browser")
+    parser = argparse.ArgumentParser(
+        prog="robocorp.browser", description="Command-line tool for robocorp.browser"
+    )
     subparsers = parser.add_subparsers(required=True)
 
     install_parser = subparsers.add_parser(
         "install",
-        help="download and install Playwright browsers",
+        help="Download and install Playwright browsers",
     )
     install_parser.set_defaults(func=_install_command)
     install_parser.add_argument(
-        "engine", type=BrowserEngine, choices=[e.value for e in BrowserEngine]
+        "engine",
+        nargs="+",
+        type=BrowserEngine,
+        choices=[e.value for e in BrowserEngine],
+        help="Browser type to install",
     )
-    install_parser.add_argument("force", action="store_true")
+    install_parser.add_argument(
+        "-f",
+        "--force",
+        action="store_true",
+        help="Force install, overwrites existing Chrome and Microsoft Edge installations",
+    )
+    install_parser.add_argument(
+        "--isolated",
+        action="store_true",
+        help="Install into holotree for caching (should only be used during post-install)",
+    )
 
     args = parser.parse_args()
     args.func(args)
@@ -24,6 +41,12 @@ def main():
 
 def _install_command(args):
     try:
-        install_browser(args.engine, force=args.force, interactive=True)
+        for engine in args.engine:
+            install_browser(
+                engine,
+                force=args.force,
+                interactive=True,
+                isolated=args.isolated,
+            )
     except (KeyboardInterrupt, InstallError):
         sys.exit(1)
