+++ browser/tests/browser_tests/test_browser_install/task_install_browser.py
@@ -7,10 +7,10 @@ def check_browser_install():
     The actual install is mocked. What the test_playwright_browser_install does
     is checking if the log output is what we expect.
     """
-    from robocorp.browser import _browser_engines
-    from robocorp.browser._browser_engines import BrowserEngine, install_browser
+    from robocorp.browser import _engines
+    from robocorp.browser._engines import BrowserEngine, install_browser
 
-    original_build_command_line = _browser_engines._build_install_command_line
+    original_build_command_line = _engines._build_install_command_line
 
     def mock_func(engine, force):
         original_cmd = original_build_command_line(engine, force)
@@ -32,9 +32,9 @@ sys.exit(1)
 """,
         ]
 
-    _browser_engines._build_install_command_line = mock_func
+    _engines._build_install_command_line = mock_func
 
     try:
         install_browser(BrowserEngine.CHROME, force=True, interactive=False)
     finally:
-        _browser_engines._build_install_command_line = original_build_command_line
+        _engines._build_install_command_line = original_build_command_line
